+++ back-end/src/main/java/com/tdd/backend/car/repository/CarRepository.java
@@ -1,4 +1,4 @@
-package com.tdd.backend.car;
+package com.tdd.backend.car.repository;
 
 import java.util.Optional;
 
@@ -6,6 +6,8 @@ import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+
+import com.tdd.backend.car.model.Car;
 
 @Repository
 public interface CarRepository extends CrudRepository<Car, Long> {
