+++ back-end/src/main/java/com/tdd/backend/post/service/DrivingService.java
@@ -6,10 +6,10 @@ import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 
-import com.tdd.backend.option.data.OptionDto;
-import com.tdd.backend.post.repository.PostRepository;
+import com.tdd.backend.car.data.OptionDto;
 import com.tdd.backend.post.data.AppointmentDto;
 import com.tdd.backend.post.data.DrivingResponse;
+import com.tdd.backend.post.repository.PostRepository;
 import com.tdd.backend.post.model.Appointment;
 import com.tdd.backend.post.model.Location;
 import com.tdd.backend.post.model.Option;
