+++ back-end/src/test/java/com/tdd/backend/post/PostControllerTest.java
@@ -16,7 +16,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.tdd.backend.option.data.OptionDto;
+import com.tdd.backend.car.data.OptionDto;
 import com.tdd.backend.post.repository.PostRepository;
 import com.tdd.backend.post.data.LocationDto;
 import com.tdd.backend.post.data.PostDto;
